+++ src/main/java/chalmers/pimp/model/canvas/layer/Raster.java
@@ -100,7 +100,7 @@ final class Raster implements IRasterLayer {
   }
 
   @Override
-  public void setRotation(double rotation) {
+  public void setRotation(int rotation) {
     layerDelegate.setRotationDegrees(rotation);
   }
 
@@ -145,7 +145,7 @@ final class Raster implements IRasterLayer {
   }
 
   @Override
-  public double getRotation() {
+  public int getRotation() {
     return layerDelegate.getRotationDegrees();
   }
